+++ ScriptedScreensHtml/PlainMarkup.cs
@@ -2115,6 +2115,8 @@ internal static partial class PlainTranslator
                 else if (part.Plan != null) lua.Emit(pad + "v_state(" + part.Plan.Var + ", " + Key(o) + ")");
                 // the elements a write makes are new ones: what listened on the old ones is gone
                 if (part == of) foreach (var made in of.Made.Where(m => m.Listens)) lua.Emit(pad + "v_unlisten(" + Q(made.Name) + ")");
+                foreach (var made in of.Made.Where(m => m.Listens))
+                    foreach (var key in made.Presses.Select(p => made.Name + " " + p).Prepend(made.Name)) lua.Emit(pad + "v_unlisten(" + Q(key) + ")");
                 Ops(o.Ops, pad);
                 foreach (var rep in o.Reps)
                     for (var k = 0; k < rep.Max; k++)